refac: events: mouse event is always reported
Before mouse events were only reported if the clicked component is
focusable.  This is unpractical because we might not want for example a
button bar get the focus if one of its buttons is clicked, but we still
want the click to be reported.  Hence the implementation now reports
always the click and iff a clicked component is focusable the focus is
moved to this component before the click is reported.