fix: report: ignore without layout wrapper
Reporting events must be ignored for components which are not part of
the layout.  This problem appeared so far only during testing:  A lines
client has components which are for some reason not part of the layout
but a Testing instance fires an event for this component.  Trying to
enable such a component to report fired event will panic since it has no
layout wrapper from which we try to retrieve the client component to
enable the component.  The later makes no sens anyway since there is no
well defined environment for such a component.  I.e. as of now as quick
fix event reporting is suppressed for such components.