fix: component: lines content replacement
So far the replacing writer implementation(s) of an environment replaced
the lines wrote and deleted the rest, i.e.:

    fmt.Fprint(e.LL(2) "line 2")
    fmt.Fprint(e.LL(1) "line 1")

contained only "line 1".  This behavior originated in the first env
writer implementation, e.g.:

    fmt.Fprint(e, "some content\nover two lines")
    fmt.Fprint(e, "overwritten by single line")

which is expected to contain only the one line "overw...".  Replacing a
component's line with different content while the remaining lines stay
as is seems in case of addressed lines the more common behavior.  Hence
when lines are addressed only the addressed line(s) are overwritten
while the rest stays as is.  A component's Reset-method may be used to
clear individual lines.