fix: events: regression: sync-chan in production
Before the event loop drained the sync-channel itself in case it was not
empty to not block the event loop.  This let to synchronization problems
in tests with nested events.  To avoid this the draining from the event
loop was removed and the testing code has now to ensure that the sync
channel gets drained otherwise no more events are reported.  In this
context I overlooked that the Events-New constructor for production also
sets the sync-channel but doesn't drain it, hence the event loop blocked
after the first write to sync.  The sync channel gets now not set if in
empty to not block the event loop.  This let to synchronization problems
in tests with nested events.  To avoid this the draining from the event
loop was removed and the testing code has now to ensure that the sync
channel gets drained otherwise no more events are reported.  In this
context I overlooked that the Events-New constructor for production also
sets the sync-channel but doesn't drain it, hence the event loop blocked
after the first write to sync.  The sync channel is now unset in
production.