refac: component: lines: writing and styling
Using the LL(line)-writer to write to a components line will set the
style of the whole line to the component's style.  Using the At(line,
cell)-writer will add a style range for given line starting with given
cell index and ending with the addition of written string's length.
Highlighting got also reimplemented.  The default highlight highlights
now the minimal area where runes are found.

TODO when the screen line is calculated tab runes '\t' should be
expanded and the styling accordingly expanded/shifted.  The same holds
true for the expansion of a line filler.  For the highlight we need an
additional feature "HighlightFilledFocusable".  For the use case to
highlight a focused line fully independent of its content.