feat: style adaption for filler/tab expansion
E.g.: "\tred"+LineFiller+"green" would in a component with width 20
expanded to "    red        green".  If we were to set a style range
from 1-4 with the bg-color red (counting the tab as one) and a style
range 6-11 with the bg-color green then these styles are shifted to 4-7
and 15-20 during the expansion process hence no matter how the component
size changes "green" will always have a green background.