refac: removed tcell from lines-package ...
... and significantly simplified the lines/testing implementations:
Event-polling respectively the event-loop is now completely gone; hence
there is no nor Events-type.  The implementation for transactional event
processing moved to the backend and got significantly better
implemented.  To make this possible events are reported from the backend
to an event lister to know when an event was processed.  Since there is
no events instance anymore the Update-event is triggered by calling the
Update-method on a component.  The move focus event is triggered by
calling the Focus-method on a component.  The register keys/runes events
are removed in favor of a Register property on the Component type which
can be used to register key/rune events during an event-callback.