refac: Update/Focus => Lines-type
After the previous refactoring the Update and Focus methods ended up
being part of the Component-API motivated by making the previous Events
type (now Lines-type) negligible since it provided only the Quit and
WaitForQuit methods.   But adding them to the Component had the
consequence that there were now two methods which could and should be
use outside of an listener callback making the API more difficult to
understand.  Now these methods are back to the Listener type and an
Env(ironment) instance of a reported event provides the Lines-instance
reporting the event to be used for Update and Focus calls.  Hence we can
say again don't use Component-methods outside an event-handler.