refac,add: globals with propagating updates
There are defaults like tab-width for tab expansion which a client might
only to deal once with if at all.  Or a client wants to change the fore-
and background color of the application.  These things can now be easily
done by using a Lines-instances Global property providing an API to
change these kind of properties and be sure that they get propagated to
all components in the layout.  Whereas it is guaranteed that a property
which has been set explicitly for a specific component will not be
change in case of a Lines-global change.

NOTE if a component c is removed from the layout by the client the
Lines' globals are changed and then c is added again to the layout its
globals will be out of sync with Lines' globals.