fix,refac: Env/Component separation; revise docs
The Env(ironment) type's job is a component lines-content while the
Component provides the rest, i.e. component aspects like features or how
a component's line behaves e.g. is it focusable.

It also turned out to be annoying that a Component.LL.By(idx) panics if
no line with given index exists, since this is not the case if a line is
accessed from the environment to write content to it Env.LL(idx).  Hence
if By's index is non-negative it always returns a line creating it when
necessary.

Env and Components.Gaps(idx) provide now a very similar API and use both
the same line-type.

Line highlighting may be full or trimmed now due to
Component.LL.Focus.Trimmed().

Stepped back from the Fmt-setter respectively the formatting property of
a component or its lines.  A mixture of convenience printers like
Component.Gaps(idx).Framed(...) and/or adding line flags like
Component.LL.By(idx).Flag(Centered) is likely to be a more fitting
solution.