refac: term.UI.Quit() became an event
So far a Quit call terminated the event-loop directly.  It seems that
this may creates race conditions under testing if a test fatales.  The
race condition occurs between the polling go routine started by a Lines
constructor and the go routine running the test.  The former is still
processing a polled event while the later calls implicitly Quit() as a
registered Cleanup function after the test fatales.   Now a Quit call
only posts a quit-event which is then polled and processed by the
polling go routine in series.