feat: component: content source
Setting a content source allows to associated arbitrary many lines with
a component while having not more but the screen lines stored in the
component.  The component source must provide a Liner implementation
from the client which is used by the component to request needed
line-prints.  In this way lines can provide its features like scrolling
or line focusing without to bother the client except for the Liner
implementation.

Note with this feature release also go a few fixes along regarding the
interaction of gaps and scrolling, line focusing/highlighting and
nesting.