rvs(features): quitable, recursive, testing
Starting to implement the 'Editable' feature it turned out that there
are two unnecessary complications in the features implementation:

- the quitable feature doesn't belong on the component-level but on the
  package level, i.e. became part of the Lines-type by adding a Quitting
  property to it.

- the _recursive feature turned out to be also quite unnecessary since
  most of the features don't make sens for nested components anyway I
  decided that features are inherited from nesting components to their
  nesting components.  A feature that should be only set for most of the
  descendants can be removed for the ones who should not have the
  feature.

Further more a couple of bugs were fixed relating to sourced components
and their line focusing/scrolling.  Unifying as well as centralizing
the test fixture components and related fixture types was extended to
sourced components testing.