feat(component): global highlighter
Until now highlighting was specified through a global highlight style
which is by default is applied if the content of line should be
highlighted.  But this is to week for more sophisticated use cases like
highlighting a line which has words in several different styles of for
highlighting two lines within the same component in different ways.  To
enable this while keeping the default behavior a global Highlighter
property was added which is a function mapping a given style to its
"highlighted" version.  Hence this property can be changed on component
level.  For example it is replaced with a Highlighter provided by a
liner.  On the other hand a globally set Highlighter (if the user for
example changes from dark to light mode) is propagated...