add: FG/BG color combination generators
There are now functions to generate all possible foreground/background
combinations for a given foreground or for a given background
respectively.  The combinations are provided as lines.Styles and the
color ranges monochrome, system8, linux, system16 and ANSI are
supported.  For other true colors the user will have to provide #rrbbgg
values since it is unpractical to name and list all true-colors.